rearranged some stuffs
